WIP: change architecture, prepare for backend integration
