+++ src/main.jsx
@@ -3,16 +3,17 @@ import ReactDOM from "react-dom/client";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import "./index.css";
 import App from "./routes/App";
-import Main from "./routes/Main";
+import Home from "./routes/Home";
+import Start from "./components/Start";
 
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <App />, // 시작 페이지
-  },
-  {
-    path: "/prompt",
-    element: <Main />, // 메인 프롬프트 페이지
+    element: <App />,
+    children: [
+      { path: "/", element: <Start /> },
+      { path: "/home", element: <Home /> },
+    ],
   },
 ]);
