+++ src/routes/Home.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Prompt from "./Prompt";
+
+const Home = () => {
+  return (
+    <div>
+      <Prompt />
+    </div>
+  );
+};
+
+export default Home;
