--- src/routes/Main.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import Prompt from "@/components/Prompt";
-
-function Main() {
-  return (
-    <main>
-      <Prompt />
-    </main>
-  );
-}
-
-export default Main;
